requirements.txt
https://download.pytorch.org/whl/cu113/torch-1.10.1%2Bcu113-cp37-cp37m-linux_x86_64.whl
transformers==4.17.0
pytorch_lightning==1.7.2
